use new ipfs-haskell lib
