--- library/Fission/IPFS/Client/Param.hs
@@ -1,10 +1,0 @@
-module Fission.IPFS.Client.Param
-  ( CID
-  , IsRecursive
-  ) where
-
-import Servant
-import Fission.Prelude
-
-type CID = QueryParam' '[Required, Strict] "arg" Text
-type IsRecursive = QueryFlag "recursive"
