--- library/Fission/IPFS/Ignored/Types.hs
@@ -1,5 +1,0 @@
-module Fission.IPFS.Ignored.Types (Ignored) where
-
-import qualified System.FilePath.Glob as Glob
-
-type Ignored = [Glob.Pattern]
