--- library/Fission/IPFS/Process/Types.hs
@@ -1,5 +1,0 @@
-module Fission.IPFS.Process.Types (Opt) where
-
-import Fission.Prelude
-
-type Opt = String
