--- library/Fission/IPFS/SparseTree.hs
@@ -1,39 +1,0 @@
-module Fission.IPFS.SparseTree
-  ( SparseTree (..)
-  , Error.Linearization (..)
-  , linearize
-  , cIDs
-  ) where
-
-import           Fission.Prelude
-import qualified Fission.Internal.UTF8 as UTF8
-import qualified Fission.IPFS.Error    as Error
-
-import           Fission.IPFS.CID.Types
-import           Fission.IPFS.Name.Types
-import           Fission.IPFS.Path.Types
-import           Fission.IPFS.SparseTree.Types
-
-linearize :: SparseTree -> Either Error.Linearization Path
-linearize = fmap Path . go
-  where
-  go :: SparseTree -> Either Error.Linearization Text
-  go = \case
-    Stub      (Name name)    -> Right <| UTF8.textShow name
-    Content   (CID _)        -> Right ""
-    Directory [(tag, value)] -> fromPath tag <$> go value
-    badDir                   -> Left <| Error.NonLinear badDir
-    where
-      fromPath tag ""   = fromKey tag
-      fromPath tag text = fromKey tag <> "/" <> text
-
-      fromKey :: Tag -> Text
-      fromKey = UTF8.stripN 1 . \case
-        Hash (CID cid)   -> cid
-        Key  (Name name) -> UTF8.textShow name
-
--- | Get all CIDs from a 'SparseTree' (all levels)
-cIDs :: (Monoid (f CID), Applicative f) => SparseTree -> f CID
-cIDs (Stub _)       = mempty
-cIDs (Content cid)  = pure cid
-cIDs (Directory kv) = foldMap cIDs kv
