--- library/Fission/IPFS/Types.hs
@@ -1,28 +1,0 @@
--- | Types related to IPFS
-module Fission.IPFS.Types
-  ( BinPath (..)
-  , CID (..)
-  , mkCID
-  , Name (..)
-  , Opt
-  , Peer (..)
-  , Path (..)
-  , SparseTree (..)
-  , Tag (..)
-  , Timeout (..)
-  , URL (..)
-  , Ignored
-  , Gateway (..)
-  ) where
-
-import Fission.IPFS.BinPath.Types
-import Fission.IPFS.CID.Types
-import Fission.IPFS.Name.Types
-import Fission.IPFS.Path.Types
-import Fission.IPFS.Peer.Types
-import Fission.IPFS.Process.Types
-import Fission.IPFS.SparseTree.Types
-import Fission.IPFS.Timeout.Types
-import Fission.IPFS.URL.Types
-import Fission.IPFS.Ignored.Types
-import Fission.IPFS.Gateway.Types
