+++ library/Fission/Internal/Orphanage/ToServerError.hs
@@ -1,0 +1,47 @@
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+
+module Fission.Internal.Orphanage.ToServerError () where
+
+import Fission.Prelude
+import Network.IPFS.ToServerError
+import Servant.Server
+
+instance ToServerError Int where
+  toServerError = \case
+    300 -> err300
+    301 -> err301
+    302 -> err302
+    303 -> err303
+    304 -> err304
+    305 -> err305
+    307 -> err307
+
+    400 -> err400
+    401 -> err401
+    402 -> err402
+    403 -> err403
+    404 -> err404
+    405 -> err405
+    406 -> err406
+    407 -> err407
+    409 -> err409
+    410 -> err410
+    411 -> err411
+    412 -> err412
+    413 -> err413
+    414 -> err414
+    415 -> err415
+    416 -> err416
+    417 -> err417
+    418 -> err418
+    422 -> err422
+
+    500 -> err500
+    501 -> err501
+    502 -> err502
+    503 -> err503
+    504 -> err504
+    505 -> err505
+
+    n -> error <| show n <> " is not an error status code"
+
