+++ library/Fission/Web.hs
@@ -13,12 +13,15 @@ import           RIO.Process (HasProcessContext)
 
 import           Data.Has
 import           Database.Selda
-import qualified Network.HTTP.Client as HTTP
+import           Data.Swagger as Swagger
 import           Servant
 
 import qualified Fission.Config     as Config
 import           Fission.User
-import qualified Fission.IPFS.Types as IPFS
+
+import           Network.IPFS.Local.Class
+import           Network.IPFS.Remote.Class
+import qualified Network.IPFS.Types as IPFS
 
 import           Fission.Internal.Orphanage.PlainText ()
 import           Fission.Internal.Orphanage.OctetStream ()
@@ -42,26 +45,24 @@ import qualified Fission.Web.Heroku            as Heroku
 -- | Top level web API type. Handled by 'server'.
 type API = Web.Swagger.API :<|> Web.API
 
--- | The actual web server for 'API'
-app
-  :: ( Has IPFS.BinPath    cfg
-     , Has IPFS.Timeout    cfg
-     , Has IPFS.URL        cfg
-     , Has IPFS.Gateway    cfg
-     , Has IPFS.Peer       cfg
-     , Has HTTP.Manager    cfg
-     , Has Web.Host        cfg
-     , Has AWS.AccessKey   cfg
-     , Has AWS.SecretKey   cfg
-     , Has AWS.ZoneID      cfg
-     , Has AWS.DomainName  cfg
-     , Has Heroku.ID       cfg
-     , Has Heroku.Password cfg
-     , HasProcessContext   cfg
-     , HasLogFunc          cfg
-     , MonadSelda     (RIO cfg)
-     )
-    => RIO cfg Application
+app ::
+  ( MonadLocalIPFS  (RIO cfg)
+  , MonadRemoteIPFS (RIO cfg)
+  , Has IPFS.Timeout     cfg
+  , Has IPFS.Gateway     cfg
+  , Has IPFS.Peer       cfg
+  , Has Web.Host         cfg
+  , Has AWS.AccessKey    cfg
+  , Has AWS.SecretKey    cfg
+  , Has AWS.ZoneID       cfg
+  , Has AWS.DomainName   cfg
+  , Has Heroku.ID        cfg
+  , Has Heroku.Password  cfg
+  , HasProcessContext    cfg
+  , HasLogFunc           cfg
+  , MonadSelda      (RIO cfg)
+  )
+  => RIO cfg Application
 app = do
   cfg     <- ask
   auth    <- mkAuth
@@ -95,23 +96,22 @@ mkAuth = do
         :. EmptyContext
 
 -- | Web handlers for the 'API'
-server
-  :: ( Has IPFS.BinPath   cfg
-     , Has IPFS.Timeout   cfg
-     , Has IPFS.URL       cfg
-     , Has IPFS.Gateway   cfg
-     , Has IPFS.Peer      cfg
-     , Has HTTP.Manager   cfg
-     , Has Web.Host       cfg
-     , Has AWS.AccessKey  cfg
-     , Has AWS.SecretKey  cfg
-     , Has AWS.ZoneID     cfg
-     , Has AWS.DomainName cfg
-     , HasProcessContext  cfg
-     , HasLogFunc         cfg
-     , MonadSelda    (RIO cfg)
-     )
-  => Web.Host
+server ::
+  ( MonadLocalIPFS  (RIO cfg)
+  , MonadRemoteIPFS (RIO cfg)
+  , Has IPFS.Timeout     cfg
+  , Has IPFS.Gateway     cfg
+  , Has IPFS.Peer      cfg
+  , Has Web.Host         cfg
+  , Has AWS.AccessKey    cfg
+  , Has AWS.SecretKey    cfg
+  , Has AWS.ZoneID       cfg
+  , Has AWS.DomainName   cfg
+  , HasProcessContext    cfg
+  , HasLogFunc           cfg
+  , MonadSelda      (RIO cfg)
+  )
+  => Swagger.Host
   -> RIOServer cfg API
 server appHost = Web.Swagger.server appHost
             :<|> IPFS.server
