+++ library/Fission/Web/DNS.hs
@@ -12,8 +12,8 @@ import           Fission.Prelude
 import           Fission.AWS.Route53
 import qualified Fission.AWS.Types   as AWS
 
-import qualified Fission.IPFS.Types as IPFS
-import           Fission.IPFS.CID.Types
+import qualified Network.IPFS.Types as IPFS
+import           Network.IPFS.CID.Types
 import           Fission.User        as User
 
 import           Fission.Web.Server
