+++ library/Fission/Web/DNS/Client.hs
@@ -5,7 +5,7 @@ import Servant.Client
 
 import           Fission.Prelude
 import qualified Fission.AWS.DomainName.Types as AWS
-import           Fission.IPFS.CID.Types
+import           Network.IPFS.CID.Types
 import qualified Fission.Web.Routes           as Routes
 
 update :: BasicAuthData -> CID -> ClientM AWS.DomainName
