+++ library/Fission/Web/IPFS/Download.hs
@@ -8,9 +8,11 @@ import           Servant
 import           Fission.Prelude
 import           Fission.Web.Server
 import qualified Fission.Web.Error        as Web.Err
-import           Fission.File.Types       as File
-import qualified Fission.IPFS.Types       as IPFS
-import qualified Fission.Storage.IPFS.Get as Storage.IPFS
+
+import           Network.IPFS.Local.Class
+import           Network.IPFS.File.Types       as File
+import qualified Network.IPFS.Types       as IPFS
+import qualified Network.IPFS.Get as IPFS
 
 type API =  PathAPI
        :<|> QueryAPI
@@ -21,31 +23,31 @@ type PathAPI = Capture "cid" IPFS.CID
 type QueryAPI = QueryParam "cid" IPFS.CID
              :> Get '[OctetStream, PlainText] File.Serialized
 
-get
-  :: ( Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , HasProcessContext cfg
-     , HasLogFunc        cfg
-     )
+get ::
+  ( MonadLocalIPFS (RIO cfg)
+  , Has IPFS.Timeout    cfg
+  , HasProcessContext   cfg
+  , HasLogFunc          cfg
+  )
   => RIOServer cfg API
 get = pathGet :<|> queryGet
 
-queryGet
-  :: ( Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , HasProcessContext cfg
-     , HasLogFunc        cfg
-     )
+queryGet ::
+  ( MonadLocalIPFS (RIO cfg) 
+  , Has IPFS.Timeout    cfg
+  , HasProcessContext   cfg
+  , HasLogFunc          cfg
+  )
   => RIOServer cfg QueryAPI
 queryGet = \case
-  Just cid -> Storage.IPFS.getFile cid >>= Web.Err.ensure
+  Just cid -> IPFS.getFile cid >>= Web.Err.ensure
   Nothing  -> throwM err404
 
-pathGet
-  :: ( Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , HasProcessContext cfg
-     , HasLogFunc        cfg
-     )
+pathGet ::
+  ( MonadLocalIPFS (RIO cfg)
+  , Has IPFS.Timeout    cfg
+  , HasProcessContext   cfg
+  , HasLogFunc          cfg
+    )
   => RIOServer cfg PathAPI
-pathGet cid = Storage.IPFS.getFile cid >>= Web.Err.ensure
+pathGet cid = IPFS.getFile cid >>= Web.Err.ensure
