+++ library/Fission/Web/IPFS/Upload.hs
@@ -5,27 +5,23 @@ module Fission.Web.IPFS.Upload
 
 import           Database.Selda
 
-import qualified Network.HTTP.Client as HTTP
 import           Servant
 
 import           Fission.Prelude
 import qualified Fission.Web.IPFS.Upload.Multipart as Multipart
 import qualified Fission.Web.IPFS.Upload.Simple    as Simple
-import qualified Fission.IPFS.Types   as IPFS
 import           Fission.Web.Server
 import           Fission.User
 
+import           Network.IPFS.Local.Class
+
 type API = Simple.API :<|> Multipart.API
 
-add
-  :: ( Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , Has HTTP.Manager  cfg
-     , Has IPFS.URL      cfg
-     , HasProcessContext cfg
-     , HasLogFunc        cfg
-     , MonadSelda   (RIO cfg)
-     )
+add ::
+  ( MonadSelda     (RIO cfg)
+  , MonadLocalIPFS (RIO cfg)
+  , HasLogFunc          cfg
+  )
   => User
   -> RIOServer cfg API
 add usr = Simple.add    usr
