+++ library/Fission/Web/IPFS/Upload/Multipart.hs
@@ -10,7 +10,6 @@ module Fission.Web.IPFS.Upload.Multipart
 import           Database.Selda
 import qualified RIO.Text as Text
 
-import qualified Network.HTTP.Client as HTTP
 import           Servant
 import           Servant.Multipart
 
@@ -20,11 +19,12 @@ import           Fission.Internal.MIME
 import           Fission.User
 import           Fission.User.CID.Mutation as User.CID
 import           Fission.Web.Server
-
-import qualified Fission.IPFS.SparseTree  as IPFS
-import qualified Fission.IPFS.Types       as IPFS
-import qualified Fission.Storage.IPFS.Add as Storage.IPFS
 import qualified Fission.Web.Error        as Web.Err
+
+import           Network.IPFS.Local.Class
+import qualified Network.IPFS.SparseTree  as IPFS
+import qualified Network.IPFS.Types       as IPFS
+import qualified Network.IPFS.Add         as IPFS
 
 type API = TextAPI :<|> JSONAPI
 
@@ -39,28 +39,20 @@ type JSONAPI = FileRequest
 type FileRequest = MultipartForm Mem (MultipartData Mem)
 type NameQuery   = QueryParam "name" IPFS.Name
 
-add
-  :: ( Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , Has HTTP.Manager  cfg
-     , Has IPFS.URL      cfg
-     , MonadSelda   (RIO cfg)
-     , HasProcessContext cfg
-     , HasLogFunc        cfg
-     )
+add ::
+  ( MonadSelda     (RIO cfg)
+  , MonadLocalIPFS (RIO cfg)
+  , HasLogFunc          cfg
+  )
   => User
   -> RIOServer cfg API
 add User { userID } = textAdd userID :<|> jsonAdd userID
 
-textAdd
-  :: ( Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , Has HTTP.Manager  cfg
-     , Has IPFS.URL      cfg
-     , HasProcessContext cfg
-     , MonadSelda   (RIO cfg)
-     , HasLogFunc        cfg
-     )
+textAdd ::
+  ( MonadSelda     (RIO cfg)
+  , MonadLocalIPFS (RIO cfg)
+  , HasLogFunc          cfg
+  )
   => ID User
   -> RIOServer cfg TextAPI
 textAdd uID form queryName = run uID form queryName <| \sparse ->
@@ -68,15 +60,11 @@ textAdd uID form queryName = run uID form queryName <| \sparse ->
     Right hash -> pure hash
     Left err   -> Web.Err.throw err
 
-jsonAdd
-  :: ( MonadSelda   (RIO cfg)
-     , Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , Has HTTP.Manager  cfg
-     , Has IPFS.URL      cfg
-     , HasProcessContext cfg
-     , HasLogFunc        cfg
-     )
+jsonAdd ::
+  ( MonadSelda     (RIO cfg)
+  , MonadLocalIPFS (RIO cfg)
+  , HasLogFunc          cfg
+  )
   => ID User
   -> RIOServer cfg JSONAPI
 jsonAdd uID form queryName = run uID form queryName pure
@@ -85,11 +73,7 @@ run
   :: ( MonadRIO          cfg m
      , MonadMask             m
      , MonadSelda            m
-     , Has HTTP.Manager  cfg
-     , Has IPFS.URL      cfg
-     , Has IPFS.BinPath  cfg
-     , Has IPFS.Timeout  cfg
-     , HasProcessContext cfg
+     , MonadLocalIPFS        m
      , HasLogFunc        cfg
      )
   => ID User
@@ -100,7 +84,7 @@ run
 run uID form qName cont = case lookupFile "file" form of
   Nothing -> throwM <| err422 { errBody = "File not processable by IPFS" }
   Just FileData { .. } ->
-    Storage.IPFS.addFile fdPayload humanName >>= \case
+    IPFS.addFile fdPayload humanName >>= \case
       Left err ->
         Web.Err.throw err
