+++ stack.yaml
@@ -49,6 +49,8 @@ extra-deps:
 - ip-1.5.1
 - raven-haskell-0.1.2.1
 - tasty-rerun-1.1.14
+- git: https://github.com/fission-suite/ipfs-haskell.git
+  commit: 61a1187c8c2c9fbc8ae6a1c58db7b75544bfb57a
 
 # Override default flag values for local packages and extra-deps
 # flags: {}
